feat: new rig modules system + resource identifier + base of a new panel editor

+++ Packages/api.nox.relay/Scripts/types/properties/InstanceRequestProperties.cs
@@ -18,7 +18,7 @@ namespace api.nox.relay.types.Properties {
 
 			foreach (var parameter in Parameters) {
 				buffer.Write(parameter.Key);
-				buffer.Write((ushort)parameter.Value.Length);
+				buffer.Write((byte)parameter.Value.Length);
 				buffer.Write(parameter.Value);
 			}
